requirements.txt
dbt-adapters==1.14.8
dbt-common==1.23.0
dbt-core==1.9.4
dbt-duckdb==1.9.4
dbt-extractor==0.6.0
dbt-semantic-interfaces==0.7.4
dbt-snowflake==1.9.4
duckdb==1.1.3
pandas==2.2.3
pandasql==0.7.3
PyYAML==6.0
streamlit==1.39.0
plotly==5.24.1
openai==1.51.0
python-dotenv==1.0.1
